requirements.txt
Django>=3.1
psycopg2-binary==2.8.3
django-cors-headers==3.5.0
django-debug-toolbar==3.1.1
django-filter==2.4.0
djangorestframework==3.12.2
drf-yasg==1.20.0
markdown==3.3.6
Pillow==8.1.0
pytz
django-leaflet